+++ app/models/export.rb
@@ -22,4 +22,12 @@ class Export < ActiveRecord::Base
     UserMailer.export_ready(self).deliver if Rails.env.production?
   end
 
+  def self.filter(filters)
+    scope = self
+    filters.each_pair do |key, value|
+      scope = scope.where(key => value) if self.column_names.include?(key.to_s) and not value.blank?
+    end
+    scope
+  end
+
 end