+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Long tables on PDFs now split across pages
 - **Variable Changes**
   - Variable Append and Prepend now display on sheet and design PDFs
+  - Domains now show up on the variables index for all variable types
 - **Gem Changes**
   - Updated to Rails 4.0.0.beta1
   - Updated to Contour 2.0.0.beta.1