+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.14.0
+## 0.14.0 (March 5, 2013)
 
 ### Enhancements
 - **Design Changes**